Added audio reactive generated tree
Added audio reactive fractal tree made in processing to get the ball rolling regarding the visuals projects.

Also, changed name of folder to generative visuals as it would be a category of it's own (coded stuff) - We can have a seperate folder for footage to be used in projection mapping.